crm_back/requirements.txt
asgiref==3.9.2
attrs==25.3.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
constantly==23.10.4
cryptography==46.0.4
daphne==4.2.1
decorator==5.2.1
Django==5.2.6
django-auditlog==3.2.1
django-cors-headers==4.9.0
django-filter==25.2
django-jazzmin==3.0.1
django-jsonfield==1.4.1
django-simple-history==3.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-extra-fields==3.7.0
drf-spectacular==0.28.0
drf-yasg==1.21.11
et_xmlfile==2.0.0
filetype==1.2.0
gunicorn==24.1.1
hyperlink==21.0.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
moviepy==2.2.1
msgpack==1.1.2
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
proglog==0.1.12
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.36.2
rest-framework-simplejwt==0.0.2
rpds-py==0.27.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
ujson==5.11.0
uritemplate==4.2.0
uvicorn[standard]==0.30.0
zope.interface==8.2
